Add files via upload
This is an ongoing to software development project. It is a python application to explore fun physics problems in nonlinear dynamics and optics. Students can learn about chaotic dynamics of pendulum, driven double pendulum, motion of charge particles in electric and magnetic fields and Double-slit experiments. No Python background is required. This platform provides graphical user interface for user to change system parameters and simulate the problem. The output is displayed in animated format that allows students to visualize chaotic trajectories. Comprehensive documentation of the problem will be provided.