Add files via upload
Can we discover an architecture where the number of two-qubit gates between distant qubits is minimised by finding the shortest path within the graph?
In the realm of graph theory, this translates to optimising the graph theoretic distance such that the paths of CNOT gates between distant qubits are the shortest possible. While this problem is NP-hard in complexity, it’s worth exploring for smaller system sizes. The code will use simulated annealing algorithm for this application.
Here’s the idea: Consider any arbitrary quantum algorithm. The key component is the number of entangling gates, which correspond to edges in graph theory (as they represent mediated coupling be- tween transmons). If we start with a randomly connected graph and minimise the distances between nodes such that paths of two-qubit operations are shortest within that graph, we anticipate reach- ing an architecture meeting the aforementioned criteria. Simulated annealing, as an inhomogeneous Markovian random walk model, is well-suited for this problem.